1: Fix a crash converting from utf8 -> utf32 2: Fix invalid truncation of utf32 char to utf8 in the lever - was causing some chars to be interpreted as 'newline' 3: Ramp up the warnings in the Xcode project 4: Fix various warnings throughout the project

+++ wolfc/platform.all/WCFileUtils.cpp
@@ -34,14 +34,14 @@ char * FileUtils::readUTF8TextFileAsUTF8String(const char * filePath, size_t & f
     }
     
     // Alloc char buffer and read
-    char * charBuffer = new char[fileSize + 1];
+    char * charBuffer = new char[static_cast<size_t>(fileSize) + 1];
     
     if (!charBuffer) {
         return nullptr;
     }
     
     if (fileSize > 0) {
-        if (std::fread(charBuffer, fileSize, 1, file) != 1) {
+        if (std::fread(charBuffer, static_cast<size_t>(fileSize), 1, file) != 1) {
             delete[] charBuffer;
             return nullptr;
         }
@@ -49,7 +49,7 @@ char * FileUtils::readUTF8TextFileAsUTF8String(const char * filePath, size_t & f
     
     // Null terminate buffer and return
     charBuffer[fileSize] = 0;
-    fileSizeInBytes = fileSize;
+    fileSizeInBytes = static_cast<size_t>(fileSize);
     return charBuffer;
 }
